Improved contents of CTL 0x1102 packet

+++ PYAS-core/aspect/compression.hpp
@@ -18,7 +18,7 @@ struct HeaderData
     float sunCenterError[2];
     int imageMinMax[2];
     float XYinterceptslope[4];
-    float CTLsolution[2];
+    double CTLsolution[2];
     float screenCenter[2];
     float screenCenterError[2];
     float fiducialX[10];
